Importing node-walker as ES6 class
The node-walker was envisioned as an extention of Ember.Object.  It
has been ported to be an ES6 class in this approach.  The general idea
is that we'll be able to share this code back to the Ember world in
the future.  For now it lives parallel to the EmberJS walker.

Note the use of the get and set methods which we've kept in order to
make the transition back to EmberJS easier in the future.