Adding jsdom dependency
It is easiest for us to add this dependency by using the package.json
file.  We also commit the package-lock.json file to ensure we're
basing ourselves on the exact same versions in the future.  In this
approach we can move fast and discuss possible upgrade paths later
down the road.