Base example of fetching triples from a context node
This example is fairly ugly.  We can easily fetch the content, but we
need to hijack its context to ensure all the prefixes are known.  We
currently steal these contexts from the RichNode which is attached to
the ContextNode.  The processing for this has been done by the
rdfa-context-scanner.

It has not been explicitly stated that this is part of a public API.
However, this content is exactly what we need to have, hence we're
using this model to correctly alter the wrapping state.