Add dumping of tag as source of node in triplestore
The source of a node can now be stored in the triplestore.  This
requires supplying the node itself, as well as the resource URI to
which it should be connected.  The result is a triple in the main
graph which indicates that the source was derived from the supplied
triple.