Automatically discover node with correct context for insertion
The previous solution required searching for the correct node to
annotate manually.  This is now detected automatically.  The contexts
are currently calculated twice, rather than only once, in order to
find the context and enrich the nodes.  It doesn't seem to be too slow
to calculate, hence we're committing it as is.