Current integration in app.js for importing a specific agenda
The import supplied here is a basic example which imports from a
specific UUID. Some extensions need to be added to ensure the correct
UUIDs exist at this time.  Furthermore, this implementation should be
adapted so it reacts to POST requests and so it consumes the UUID to
import from the request.