Adapt importAgendaFromDoc to new model
The previous model contained slightly different content.  Furthermore,
we're now not storing the original document from which the content was
derived as that seems to cause hickups when we store it in the graph.
A differently constructed query would solve this, but we'll tackle
that issue when we're also importing the decisions.