Only create VersionedAgenda if none exists
We should only create a VersionedAgenda if no VersionedAgenda exists
for this document.  The signatures and publications are already
separated from this object to indicate its publication status.

Our current approach is to search for the agenda, and create one if it
doesn't exist.  We return the URI of the VersionedAgenda without any
indication of where it came from.